+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed an issue with the initial annual interest rate in the _FIRE_ calculator
+- Fixed the state handling in the currency selector
 - Fixed the deletion of an asset profile with symbol profile overrides in the asset profile details dialog of the admin control
 
 ## 2.83.0 - 2024-05-30